pyproject.toml
[project]
name = "object-storage"
version = "0.1.0"
description = "Self-hosted object storage API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn[standard]>=0.35.0",
    "psycopg2-binary>=2.9.10",
    "sqlmodel>=0.0.24",
    "asyncpg>=0.30.0",
]
